requirements_test.txt
pytest>=7.4
pytest-asyncio>=0.23
pytest-homeassistant-custom-component>=0.13
aioresponses>=0.7
